Code refactor with RateFormModal.
