+++ js/src/components/shipping-rate-section/estimated-shipping-rates-card/edit-rate-form-modal.js
@@ -3,16 +3,12 @@
  */
 import { Button } from '@wordpress/components';
 import { __ } from '@wordpress/i18n';
-import { Form } from '@woocommerce/components';
 import { noop } from 'lodash';
 
 /**
  * Internal dependencies
  */
-import AppModal from '.~/components/app-modal';
-import AppInputPriceControl from '.~/components/app-input-price-control';
-import VerticalGapLayout from '.~/components/vertical-gap-layout';
-import AppCountrySelect from '.~/components/app-country-select';
+import RateFormModal from './rate-form-modal.js';
 
 /**
  * @typedef { import(".~/data/actions").CountryCode } CountryCode
@@ -25,119 +21,36 @@ import AppCountrySelect from '.~/components/app-country-select';
  * @param {Object} props
  * @param {Array<CountryCode>} props.countryOptions Array of country codes, to be used as options in AppCountrySelect.
  * @param {ShippingRateGroup} props.initialValues Initial values for the form.
- * @param {(newGroup: ShippingRateGroup) => void} props.onSubmit Called when the shipping rate group is submitted.
+ * @param {(values: ShippingRateGroup) => void} props.onSubmit Called when the shipping rate group is submitted.
  * @param {() => void} props.onDelete Called when users clicked on the Delete button.
  * @param {() => void} props.onRequestClose Callback to close the modal.
  */
-const EditRateFormModal = ( {
-	countryOptions,
-	initialValues,
-	onDelete = noop,
-	onSubmit = noop,
-	onRequestClose = noop,
-} ) => {
+const EditRateFormModal = ( props ) => {
+	const { onDelete = noop, onRequestClose = noop } = props;
+
 	const handleDeleteClick = () => {
 		onRequestClose();
 		onDelete();
 	};
 
-	/**
-	 * @param {ShippingRateGroup} values
-	 */
-	const handleValidate = ( values ) => {
-		const errors = {};
-
-		if ( values.countries.length === 0 ) {
-			errors.countries = __(
-				'Please specify at least one country.',
-				'google-listings-and-ads'
-			);
-		}
-
-		if ( values.rate < 0 ) {
-			errors.rate = __(
-				'The estimated shipping rate cannot be less than 0.',
-				'google-listings-and-ads'
-			);
-		}
-
-		return errors;
-	};
-
-	/**
-	 * @param {ShippingRateGroup} newGroup
-	 */
-	const handleSubmitCallback = ( newGroup ) => {
-		onRequestClose();
-		onSubmit( newGroup );
-	};
-
 	return (
-		<Form
-			initialValues={ initialValues }
-			validate={ handleValidate }
-			onSubmit={ handleSubmitCallback }
-		>
-			{ ( formProps ) => {
-				const {
-					getInputProps,
-					values,
-					isValidForm,
-					handleSubmit,
-				} = formProps;
-
-				return (
-					<AppModal
-						title={ __(
-							'Estimate a shipping rate',
-							'google-listings-and-ads'
-						) }
-						buttons={ [
-							<Button
-								key="delete"
-								isTertiary
-								isDestructive
-								onClick={ handleDeleteClick }
-							>
-								{ __( 'Delete', 'google-listings-and-ads' ) }
-							</Button>,
-							<Button
-								key="save"
-								isPrimary
-								disabled={ ! isValidForm }
-								onClick={ handleSubmit }
-							>
-								{ __(
-									'Update shipping rate',
-									'google-listings-and-ads'
-								) }
-							</Button>,
-						] }
-						onRequestClose={ onRequestClose }
-					>
-						<VerticalGapLayout>
-							<AppCountrySelect
-								label={ __(
-									'If customer is in',
-									'google-listings-and-ads'
-								) }
-								options={ countryOptions }
-								multiple
-								{ ...getInputProps( 'countries' ) }
-							/>
-							<AppInputPriceControl
-								label={ __(
-									'Then the estimated shipping rate displayed in the product listing is',
-									'google-listings-and-ads'
-								) }
-								suffix={ values.currency }
-								{ ...getInputProps( 'rate' ) }
-							/>
-						</VerticalGapLayout>
-					</AppModal>
-				);
-			} }
-		</Form>
+		<RateFormModal
+			{ ...props }
+			additionalButtons={ [
+				<Button
+					key="delete"
+					isTertiary
+					isDestructive
+					onClick={ handleDeleteClick }
+				>
+					{ __( 'Delete', 'google-listings-and-ads' ) }
+				</Button>,
+			] }
+			submitButtonChildren={ __(
+				'Update shipping rate',
+				'google-listings-and-ads'
+			) }
+		/>
 	);
 };
