+++ js/src/components/shipping-rate-section/estimated-shipping-rates-card/rate-form-modal.js
@@ -1,0 +1,133 @@
+/**
+ * External dependencies
+ */
+import { Button } from '@wordpress/components';
+import { __ } from '@wordpress/i18n';
+import { Form } from '@woocommerce/components';
+import { noop } from 'lodash';
+
+/**
+ * Internal dependencies
+ */
+import AppModal from '.~/components/app-modal';
+import AppInputPriceControl from '.~/components/app-input-price-control/index.js';
+import VerticalGapLayout from '.~/components/vertical-gap-layout';
+import AppCountrySelect from '.~/components/app-country-select';
+
+/**
+ * @typedef { import(".~/data/actions").CountryCode } CountryCode
+ * @typedef { import("./typedefs.js").ShippingRateGroup } ShippingRateGroup
+ */
+
+/**
+ * Base rate form modal.
+ *
+ * This is used by AddRateFormModal and EditRateFormModal.
+ *
+ * @param {Object} props
+ * @param {JSX.Element} props.submitButtonChildren Children for the submit button.
+ * @param {Array<Button>} [props.additionalButtons=[]] Additional buttons to be rendered in `AppModal`'s `buttons` prop, before the submit button.
+ * @param {Array<CountryCode>} props.countryOptions Array of country codes, to be used as options in AppCountrySelect.
+ * @param {ShippingRateGroup} props.initialValues Initial values for the form.
+ * @param {() => void} props.onRequestClose Callback to close the modal.
+ * @param {(values: ShippingRateGroup) => void} props.onSubmit Called with submitted value.
+ */
+const RateFormModal = ( {
+	additionalButtons = [],
+	submitButtonChildren,
+	countryOptions,
+	initialValues,
+	onSubmit = noop,
+	onRequestClose = noop,
+} ) => {
+	/**
+	 * @param {ShippingRateGroup} values
+	 */
+	const handleValidate = ( values ) => {
+		const errors = {};
+
+		if ( values.countries.length === 0 ) {
+			errors.countries = __(
+				'Please specify at least one country.',
+				'google-listings-and-ads'
+			);
+		}
+
+		if ( values.rate < 0 ) {
+			errors.rate = __(
+				'The estimated shipping rate cannot be less than 0.',
+				'google-listings-and-ads'
+			);
+		}
+
+		return errors;
+	};
+
+	/**
+	 * @param {ShippingRateGroup} values
+	 */
+	const handleSubmitCallback = ( values ) => {
+		onRequestClose();
+		onSubmit( values );
+	};
+
+	return (
+		<Form
+			initialValues={ initialValues }
+			validate={ handleValidate }
+			onSubmit={ handleSubmitCallback }
+		>
+			{ ( formProps ) => {
+				const {
+					getInputProps,
+					values,
+					isValidForm,
+					handleSubmit,
+				} = formProps;
+
+				return (
+					<AppModal
+						title={ __(
+							'Estimate a shipping rate',
+							'google-listings-and-ads'
+						) }
+						buttons={ [
+							...additionalButtons,
+							<Button
+								key="submit"
+								isPrimary
+								disabled={ ! isValidForm }
+								onClick={ handleSubmit }
+							>
+								{ submitButtonChildren }
+							</Button>,
+						] }
+						onRequestClose={ onRequestClose }
+					>
+						<VerticalGapLayout>
+							<AppCountrySelect
+								label={ __(
+									'If customer is in',
+									'google-listings-and-ads'
+								) }
+								options={ countryOptions }
+								multiple
+								{ ...getInputProps( 'countries' ) }
+							/>
+							<AppInputPriceControl
+								label={ __(
+									'Then the estimated shipping rate displayed in the product listing is',
+									'google-listings-and-ads'
+								) }
+								suffix={ values.currency }
+								{ ...getInputProps( 'rate' ) }
+							/>
+						</VerticalGapLayout>
+					</AppModal>
+				);
+			} }
+		</Form>
+	);
+};
+
+export default RateFormModal;
